rename zod config to schema
